Fix compiler warnings about signedness in src/ansi_extras.h

+++ src/ansi_extras.h
@@ -67,11 +67,11 @@ static inline char* strdupn(const char* s, size_t len)
 	return ret;
 }
 
-static inline void print_short(char* buf, int max_len, const char* src)
+static inline void print_short(char* buf, size_t max_len, const char* src)
 {
-	int rem_len = max_len - 3 - 1;
-	int suff_len = rem_len/2;
-	int pref_len = rem_len-suff_len;
+	size_t rem_len = max_len - 3 - 1;
+	size_t suff_len = rem_len/2;
+	size_t pref_len = rem_len-suff_len;
 
 	if(max_len < 4 || strlen(src) < max_len) {
 		strncpy(buf, src, max_len);
@@ -79,11 +79,11 @@ static inline void print_short(char* buf, int max_len, const char* src)
 	}
 
 	if(max_len < 6) {
-		snprintf(buf, max_len, "%.*s...", max_len-3, src);
+		snprintf(buf, max_len, "%.*s...", (int)(max_len-3), src);
 		return;
 	}
 
-	snprintf(buf, max_len, "%.*s...%.*s", pref_len, src, suff_len, &src[strlen(src)-suff_len]);
+	snprintf(buf, max_len, "%.*s...%.*s", (int)pref_len, src, (int)suff_len, &src[strlen(src)-suff_len]);
 }
 
 static inline int isodigit(char c)
